pyproject.toml
[tool.poetry]
name = "mini-print-stream-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
ultralytics = "^8.2.18"
flask = "^3.0.3"
websockets = "^12.0"
pydantic = "^2.7.1"
dill = "^0.3.8"
python-statemachine = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
